# packages/contracts/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Job Getter API",
    "version": "0.0.0",
    "description": "Generated from the TypeBox contracts in packages/contracts by packages/contracts/scripts/generate.ts. Do not edit by hand. JSON snake_case, ISO-8601 UTC timestamps, UUID identifiers. The API derives the workspace from the authenticated principal and never trusts a client workspace_id."
  },
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jg_session",
        "description": "HttpOnly session cookie issued by POST /auth/login."
      },
      "deviceToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Scoped device token for the extension or local runner; expires after 30 days."
      },
      "workerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Operator worker credential for the internal task protocol only."
      }
    },
    "schemas": {
...
      "ContactValue": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "full_name",
          "email"
        ],
        "properties": {
          "full_name": {
            "minLength": 1,
            "maxLength": 200,
            "type": "string"
          },
          "email": {
            "minLength": 3,
            "maxLength": 320,
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "city": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "country": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "links": {
            "maxItems": 12,
            "type": "array",
            "items": {
              "additionalProperties": false,
              "type": "object",
              "required": [
                "label",
                "url"
              ],
              "properties": {
                "label": {
                  "minLength": 1,
                  "maxLength": 60,
                  "type": "string"
                },
                "url": {
                  "minLength": 1,
                  "maxLength": 500,
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "request_id"
            ],
            "properties": {
              "code": {
                "anyOf": [
                  {
                    "const": "VALIDATION_ERROR",
                    "type": "string"
                  },
                  {
                    "const": "MALFORMED_REQUEST",
                    "type": "string"
                  },
                  {
                    "const": "UNAUTHENTICATED",
                    "type": "string"
                  },
                  {
                    "const": "FORBIDDEN",
                    "type": "string"
                  },
                  {
                    "const": "NOT_FOUND",
                    "type": "string"
                  },
                  {
                    "const": "CONFLICT",
                    "type": "string"
                  },
                  {
                    "const": "STALE_REVISION",
                    "type": "string"
                  },
                  {
                    "const": "IDEMPOTENCY_MISMATCH",
                    "type": "string"
                  },
                  {
                    "const": "PAYLOAD_TOO_LARGE",
                    "type": "string"
                  },
                  {
                    "const": "UNPROCESSABLE",
                    "type": "string"
                  },
                  {
                    "const": "QUOTA_EXCEEDED",
                    "type": "string"
                  },
                  {
                    "const": "SETUP_CLOSED",
                    "type": "string"
                  },
                  {
                    "const": "BUDGET_EXHAUSTED",
                    "type": "string"
                  },
                  {
                    "const": "PROVIDER_UNAVAILABLE",
                    "type": "string"
                  },
                  {
                    "const": "INTERNAL_ERROR",
                    "type": "string"
                  }
                ]
              },
              "message": {
                "type": "string"
              },
              "fields": {
                "type": "object",
                "patternProperties": {
                  "^(.*)$": {
                    "type": "string"
                  }
                }
              },
              "request_id": {
                "format": "uuid",
                "type": "string"
              }
            }
          }
        }
      },
...
      "PreferencesPutRequest": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_revision",
          "config"
        ],
        "properties": {
          "expected_revision": {
            "minimum": 1,
            "type": "integer"
          },
          "config": {
            "$ref": "#/components/schemas/Preferences"
          }
        }
      },
      "PreferencesView": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revision",
          "config",
          "updated_at"
        ],
        "properties": {
          "revision": {
            "minimum": 1,
            "type": "integer"
          },
          "config": {
            "$ref": "#/components/schemas/Preferences"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        }
      },
      "Profile": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "revision",
          "confirmed_revision",
          "contact",
          "locale",
          "facts",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "revision": {
            "minimum": 1,
            "type": "integer"
          },
          "confirmed_revision": {
            "anyOf": [
              {
                "minimum": 1,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "contact": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContactValue"
              },
              {
                "type": "null"
              }
            ]
          },
          "locale": {
            "anyOf": [
              {
                "const": "en",
                "type": "string"
              },
              {
                "const": "es",
                "type": "string"
              }
            ]
          },
          "facts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfileFact"
            }
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        }
      },
      "ProfileFact": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "kind",
          "value",
          "source_file_id",
          "source_excerpt",
          "confirmed",
          "revision",
          "supersedes_id",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "kind": {
            "anyOf": [
              {
                "const": "contact",
                "type": "string"
              },
              {
                "const": "summary",
                "type": "string"
              },
              {
                "const": "experience",
                "type": "string"
              },
              {
                "const": "education",
                "type": "string"
              },
              {
                "const": "skill",
                "type": "string"
              },
              {
                "const": "language",
                "type": "string"
              },
              {
                "const": "authorization",
                "type": "string"
              },
              {
                "const": "project",
                "type": "string"
              },
              {
                "const": "certification",
                "type": "string"
              }
            ]
          },
          "value": {},
          "source_file_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "source_excerpt": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "confirmed": {
            "type": "boolean"
          },
          "revision": {
            "minimum": 1,
            "type": "integer"
          },
          "supersedes_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        }
      },
...
      "ProfilePatchRequest": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_revision",
          "changes"
        ],
        "properties": {
          "expected_revision": {
            "minimum": 1,
            "type": "integer"
          },
          "contact": {
            "$ref": "#/components/schemas/ContactValue"
          },
          "locale": {
            "anyOf": [
              {
                "const": "en",
                "type": "string"
              },
              {
                "const": "es",
                "type": "string"
              }
            ]
          },
          "changes": {
            "maxItems": 200,
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "op",
                    "kind",
                    "value",
                    "confirmed"
                  ],
                  "properties": {
                    "op": {
                      "const": "upsert",
                      "type": "string"
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "kind": {
                      "anyOf": [
                        {
                          "const": "contact",
                          "type": "string"
                        },
                        {
                          "const": "summary",
                          "type": "string"
                        },
                        {
                          "const": "experience",
                          "type": "string"
                        },
                        {
                          "const": "education",
                          "type": "string"
                        },
                        {
                          "const": "skill",
                          "type": "string"
                        },
                        {
                          "const": "language",
                          "type": "string"
                        },
                        {
                          "const": "authorization",
                          "type": "string"
                        },
                        {
                          "const": "project",
                          "type": "string"
                        },
                        {
                          "const": "certification",
                          "type": "string"
                        }
                      ]
                    },
                    "value": {},
                    "confirmed": {
                      "type": "boolean"
                    },
                    "source_excerpt": {
                      "anyOf": [
                        {
                          "maxLength": 2000,
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "source_file_id": {
                      "anyOf": [
                        {
                          "format": "uuid",
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                },
                {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "op",
                    "id"
                  ],
                  "properties": {
                    "op": {
                      "const": "delete",
                      "type": "string"
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/profile": {
      "get": {
        "operationId": "getProfile",
        "summary": "Profile with facts and the current revision.",
        "tags": [
          "profile"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile with facts and the current revision.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request body, query or header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated: no valid session, device token or worker credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted for this workspace resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Absent or not accessible from the authenticated workspace.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: stale revision, duplicate command or idempotency key reuse with a different body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Payload exceeds the configured upload limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Structurally valid but invalid domain input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Quota or rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "patchProfile",
        "summary": "Apply fact changes against an expected revision.",
        "tags": [
          "profile"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Apply fact changes against an expected revision.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request body, query or header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated: no valid session, device token or worker credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted for this workspace resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Absent or not accessible from the authenticated workspace.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: stale revision, duplicate command or idempotency key reuse with a different body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Payload exceeds the configured upload limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Structurally valid but invalid domain input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Quota or rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "description": "Anti-CSRF token for cookie-authenticated mutations; origin is verified too.",
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfilePatchRequest"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/preferences": {
      "get": {
        "operationId": "getPreferences",
        "summary": "Validated preference config and revision.",
        "tags": [
          "preferences"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Validated preference config and revision.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesView"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request body, query or header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated: no valid session, device token or worker credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted for this workspace resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Absent or not accessible from the authenticated workspace.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: stale revision, duplicate command or idempotency key reuse with a different body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Payload exceeds the configured upload limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Structurally valid but invalid domain input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Quota or rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "putPreferences",
        "summary": "Replace preferences; unknown keys are rejected.",
        "tags": [
          "preferences"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Replace preferences; unknown keys are rejected.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesView"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request body, query or header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated: no valid session, device token or worker credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted for this workspace resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Absent or not accessible from the authenticated workspace.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: stale revision, duplicate command or idempotency key reuse with a different body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Payload exceeds the configured upload limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Structurally valid but invalid domain input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Quota or rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "description": "Anti-CSRF token for cookie-authenticated mutations; origin is verified too.",
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferencesPutRequest"
              }
            }
          }
        }
      }
    },
...
  }
}